feat: create default cache when configuring profile
